+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0
 
 - Handle -h and -v options properly.
+- added `linker-args` configuration option.
 
 ## 0.1.0