+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+
+- Renamed `clean` task to `cleanC` (to prevent conflict with other Dartle tools) and documented it.
+
 ## 0.2.0
 
 - Handle -h and -v options properly.